LayoutTests/http/tests/contentextensions/prefetch-blocked.html

This page prefetches a Document.<br>
The prefetch should be blocked by the content extension filter<br>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
}
</script>
<link rel="prefetch" href="resources/should-not-load.html">
