LayoutTests/http/tests/contentextensions/prefetch-blocked-expected.txt

CONSOLE MESSAGE: line 8: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/resources/should-not-load.html from loading a resource from http://127.0.0.1:8000/contentextensions/resources/should-not-load.html
This page prefetches a Document.
The prefetch should be blocked by the content extension filter
